style: add eslint; apply linting rules
