+++ package.json
@@ -11,7 +11,7 @@
     "build:ts": "tsc -p ./",
     "test": "npm run lint",
     "startmon": "ts-node-dev --respawn ./src/index.ts",
-    "lint": "tslint --project . src/**/*.ts",
+    "lint": "eslint --ext .js,.ts --ignore-path .gitignore src/",
     "lint:fix": "npm run lint --fix",
     "watch": "tsc -w -p ./",
     "clean": "rimraf ./build && mkdir build",
@@ -43,13 +43,20 @@
     "@types/imagemin-gifsicle": "^5.2.0",
     "@types/imagemin-jpegtran": "^5.0.0",
     "@types/moment-duration-format": "^2.2.2",
-    "@types/node": "^14.6.0",
+    "@types/node": "^14.6.3",
     "@types/ws": "^7.2.6",
+    "@typescript-eslint/eslint-plugin": "^4.0.1",
+    "@typescript-eslint/parser": "^4.0.1",
     "cz-conventional-changelog": "^3.3.0",
+    "eslint": "^7.8.1",
+    "eslint-config-standard": "^14.1.1",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.1",
     "husky": "^4.2.5",
     "rimraf": "^3.0.2",
-    "ts-node-dev": "^1.0.0-pre.59",
-    "tslint": "^6.1.3",
+    "ts-node-dev": "^1.0.0-pre.62",
     "typescript": "^4.0.2"
   },
   "repository": {
