--- tslint.json
@@ -1,6 +1,0 @@
-{
-  "extends": "@tmuniversal/ts/lint",
-  "rules": {
-    "no-floating-promises": false
-  }
-}
